builtin reuse
